edge-agent/edge-agent/requirements.txt
anthropic>=0.18.0
httpx>=0.25.0
structlog>=23.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
py-clob-client>=0.15.0
cryptography>=41.0.0
web3>=6.0.0
